Player with Grid Movement
Made a grid system that the player moves on. Added a camera that tracks the player with cinemachine. (the movement system might have to be redone as it is kind of hard coded and might not be able to be added to the enemies easily) apart from that, its going pretty well!